rename:  패키지 이름 변경(score -> winningstatus)
