+++ src/main/java/chess/domain/winningstatus/Score.java
@@ -1,4 +1,4 @@
-package chess.domain.score;
+package chess.domain.winningstatus;
 
 import chess.domain.piece.SquareState;
